Pull projects out into shared code
