+++ packages/excavator-projects/index.ts
@@ -1,0 +1,33 @@
+import type { ExcavatorProject } from "./type";
+
+import { AUTUMNATON } from "./projects/autumnaton";
+import { BIRD_A_DAY } from "./projects/birdADay";
+import { COAT_OF_PAINT } from "./projects/coatOfPaint";
+import { DROP_CON_SNOWGLOBE } from "./projects/dropConSnowglobe";
+import { DROP_MIXED_EVERYTHING } from "./projects/dropMixedEverything";
+import { DROP_MR_CHEENGS } from "./projects/dropMrCheengs";
+import { DROP_MR_SCREEGES } from "./projects/dropMrScreeges";
+import { GENIE } from "./projects/genie";
+import { HOOKAH } from "./projects/hookah";
+import { JUICE_BAR } from "./projects/juiceBar";
+import { MONSTER_PARTS } from "./projects/monsterParts";
+import { MUMMING_TRUNK } from "./projects/mummingTrunk";
+import { OUT_OF_ORDER } from "./projects/outOfOrder";
+
+export { ExcavatorProject };
+
+export const projects: ExcavatorProject[] = [
+  AUTUMNATON,
+  BIRD_A_DAY,
+  COAT_OF_PAINT,
+  GENIE,
+  HOOKAH,
+  JUICE_BAR,
+  DROP_CON_SNOWGLOBE,
+  DROP_MIXED_EVERYTHING,
+  DROP_MR_CHEENGS,
+  DROP_MR_SCREEGES,
+  MONSTER_PARTS,
+  MUMMING_TRUNK,
+  OUT_OF_ORDER,
+];
