+++ packages/excavator-projects/projects/dropMrScreeges.ts
@@ -1,23 +1,23 @@
 /**
  * @author rinn
- * Track drops from the Mr. Cheeng's Spectacles
+ * Track drops from the Mr. Screege's Spectacles
  */
 import { currentRound, equippedAmount, Item } from "kolmafia";
 import { $item } from "libram";
 
 import { ExcavatorProject } from "../type";
-import { toNormalisedString } from "../utils/game";
+import { toNormalisedString } from "../utils";
 
-export const DROP_MR_CHEENGS: ExcavatorProject = {
-  name: "Mr. Cheeng's Spectacles",
+export const DROP_MR_SCREEGES: ExcavatorProject = {
+  name: "Mr. Screege's Spectacles",
   hooks: {
     COMBAT_ROUND: (encounter: string, page: string) => {
       // Must be end of battle
       if (currentRound() !== 0) return null;
-      // Must be wearing Mr. Cheeng's spectacles
-      if (equippedAmount($item`Mr. Cheeng's spectacles`) < 1) return null;
+      // Must be wearing Mr. Screege's spectacles
+      if (equippedAmount($item`Mr. Screege's spectacles`) < 1) return null;
       const result = page.match(
-        /You see a weird thing out of the corner of your eye, and you grab it.\s+Far out, man!.*?You acquire an item: <b>(.*?)<\/b>/,
+        /You notice something valuable hidden .*?You acquire an item: <b>(.*?)<\/b>/,
       );
       if (!result) return null;
       const item = Item.get(result[1]);
