--- packages/excavator-script/src/projects/index.ts
@@ -1,31 +1,0 @@
-import type { ExcavatorProject } from "../type";
-
-import { AUTUMNATON } from "./autumnaton";
-import { BIRD_A_DAY } from "./birdADay";
-import { COAT_OF_PAINT } from "./coatOfPaint";
-import { DROP_CON_SNOWGLOBE } from "./dropConSnowglobe";
-import { DROP_MIXED_EVERYTHING } from "./dropMixedEverything";
-import { DROP_MR_CHEENGS } from "./dropMrCheengs";
-import { DROP_MR_SCREEGES } from "./dropMrScreeges";
-import { GENIE } from "./genie";
-import { HOOKAH } from "./hookah";
-import { JUICE_BAR } from "./juiceBar";
-import { MONSTER_PARTS } from "./monsterParts";
-import { MUMMING_TRUNK } from "./mummingTrunk";
-import { OUT_OF_ORDER } from "./outOfOrder";
-
-export const projects: ExcavatorProject[] = [
-  AUTUMNATON,
-  BIRD_A_DAY,
-  COAT_OF_PAINT,
-  GENIE,
-  HOOKAH,
-  JUICE_BAR,
-  DROP_CON_SNOWGLOBE,
-  DROP_MIXED_EVERYTHING,
-  DROP_MR_CHEENGS,
-  DROP_MR_SCREEGES,
-  MONSTER_PARTS,
-  MUMMING_TRUNK,
-  OUT_OF_ORDER,
-];
